refactor(api): reorganize commonwell-cert
Ref #227
