+++ packages/packages/commonwell-sdk/package.json
@@ -20,7 +20,7 @@
   "repository": {
     "url": "https://github.com/metriport/metriport.git",
     "type": "git",
-    "directory": "pkgs/packages/commonwell-sdk"
+    "directory": "packages/packages/commonwell-sdk"
   },
   "scripts": {
     "tsc": "tsc",
